:recycle: refactor: Course PK 변경에 의한 findTop10PopularChallenges() 메서드 수정
