+++ src/main/java/com/saisai/domain/challenge/repository/ChallengeParticipantRepositoryCustom.java
@@ -1,6 +1,6 @@
 package com.saisai.domain.challenge.repository;
 
-import com.saisai.domain.challenge.dto.response.ChallengeInfoProjection;
+import com.saisai.domain.challenge.dto.projection.ChallengeInfoProjection;
 import java.util.List;
 
 public interface ChallengeParticipantRepositoryCustom {
