+++ src/main/java/com/saisai/domain/challenge/repository/ChallengeParticipantRepositoryImpl.java
@@ -2,11 +2,13 @@ package com.saisai.domain.challenge.repository;
 
 import static com.saisai.domain.challenge.entity.QChallenge.challenge;
 import static com.saisai.domain.challenge.entity.QChallengeParticipant.challengeParticipant;
+import static com.saisai.domain.course.entity.QCourse.course;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import com.saisai.domain.challenge.dto.response.ChallengeInfoProjection;
-import com.saisai.domain.challenge.dto.response.QChallengeInfoProjection;
+import com.saisai.domain.challenge.dto.projection.ChallengeInfoProjection;
+import com.saisai.domain.challenge.dto.projection.QChallengeInfoProjection;
 import com.saisai.domain.challenge.entity.ChallengeStatus;
+import com.saisai.domain.course.repository.CourseRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
@@ -16,27 +18,23 @@ import org.springframework.stereotype.Repository;
 public class ChallengeParticipantRepositoryImpl implements ChallengeParticipantRepositoryCustom {
 
     private final JPAQueryFactory queryFactory;
+    private final CourseRepository courseRepository;
 
     @Override
     public List<ChallengeInfoProjection> findTop10PopularChallenges() {
         return queryFactory
             .select(new QChallengeInfoProjection(
                 challenge.id,
-                challenge.courseName,
+                course.id,
                 challenge.status,
                 challenge.endedAt,
-                challengeParticipant.user.id.count().as("participantCount")
+                challengeParticipant.count().as("participantCount")
             ))
             .from(challengeParticipant)
-            .join(challenge).on(challengeParticipant.challenge.id.eq(challenge.id))
+            .join(challengeParticipant.challenge, challenge)
             .where(challenge.status.eq(ChallengeStatus.ONGOING))
-            .groupBy(
-                challenge.id,
-                challenge.courseName,
-                challenge.status,
-                challenge.endedAt
-            )
-            .orderBy(challengeParticipant.user.id.count().desc())
+            .groupBy(challenge.id)
+            .orderBy(challengeParticipant.count().desc())
             .limit(10)
             .fetch();
     }
